realtime_price_agent/legacy/supply-chain-agents/agents/demand_forecasting/requirements.txt
agent-framework
agent-framework-azure-ai>=1.0.0
azure-identity>=1.15.0
azure-ai-openai>=1.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0